chore(gui): Minor refactor
