+++ src/main/java/it/units/sdm/quoridor/gui/BorderManager.java
@@ -1,4 +1,4 @@
-package it.units.sdm.quoridor.GUI;
+package it.units.sdm.quoridor.gui;
 
 import javax.swing.*;
 import javax.swing.border.AbstractBorder;
