+++ src/main/java/it/units/sdm/quoridor/gui/buttons/MoveButton.java
@@ -1,8 +1,8 @@
-package it.units.sdm.quoridor.GUI.buttons;
+package it.units.sdm.quoridor.gui.buttons;
 
-import it.units.sdm.quoridor.GUI.DialogManager;
-import it.units.sdm.quoridor.GUI.GUIConstants;
-import it.units.sdm.quoridor.GUI.GameBoardGUI;
+import it.units.sdm.quoridor.gui.DialogManager;
+import it.units.sdm.quoridor.gui.GUIConstants;
+import it.units.sdm.quoridor.gui.GameBoardGUI;
 
 import javax.swing.*;
