+++ src/main/java/it/units/sdm/quoridor/gui/panels/ActionsPanelComponent.java
@@ -1,11 +1,11 @@
-package it.units.sdm.quoridor.GUI.panels;
+package it.units.sdm.quoridor.gui.panels;
 
-import it.units.sdm.quoridor.GUI.DialogManager;
-import it.units.sdm.quoridor.GUI.GUIConstants;
-import it.units.sdm.quoridor.GUI.GameBoardGUI;
-import it.units.sdm.quoridor.GUI.GameController;
-import it.units.sdm.quoridor.GUI.buttons.MoveButton;
-import it.units.sdm.quoridor.GUI.buttons.PlaceWallButton;
+import it.units.sdm.quoridor.gui.DialogManager;
+import it.units.sdm.quoridor.gui.GUIConstants;
+import it.units.sdm.quoridor.gui.GameBoardGUI;
+import it.units.sdm.quoridor.gui.GameController;
+import it.units.sdm.quoridor.gui.buttons.MoveButton;
+import it.units.sdm.quoridor.gui.buttons.PlaceWallButton;
 
 import javax.swing.*;
 import java.awt.*;
