+++ src/main/java/it/units/sdm/quoridor/gui/panels/PanelComponent.java
@@ -1,4 +1,4 @@
-package it.units.sdm.quoridor.GUI.panels;
+package it.units.sdm.quoridor.gui.panels;
 
 import javax.swing.JPanel;
