+++ src/main/java/it/units/sdm/quoridor/gui/panels/PlayerPanelsComponent.java
@@ -1,8 +1,8 @@
-package it.units.sdm.quoridor.GUI.panels;
+package it.units.sdm.quoridor.gui.panels;
 
-import it.units.sdm.quoridor.GUI.GUIConstants;
-import it.units.sdm.quoridor.GUI.GameBoardGUI;
-import it.units.sdm.quoridor.GUI.GameController;
+import it.units.sdm.quoridor.gui.GUIConstants;
+import it.units.sdm.quoridor.gui.GameBoardGUI;
+import it.units.sdm.quoridor.gui.GameController;
 import it.units.sdm.quoridor.model.AbstractPawn;
 
 import javax.swing.*;
