+++ src/main/java/it/units/sdm/quoridor/gui/panels/WallDirectionsPanelComponent.java
@@ -1,10 +1,10 @@
-package it.units.sdm.quoridor.GUI.panels;
+package it.units.sdm.quoridor.gui.panels;
 
-import it.units.sdm.quoridor.GUI.GUIConstants;
-import it.units.sdm.quoridor.GUI.GameBoardGUI;
-import it.units.sdm.quoridor.GUI.GameController;
-import it.units.sdm.quoridor.GUI.buttons.CancelButton;
-import it.units.sdm.quoridor.GUI.buttons.OrientationButton;
+import it.units.sdm.quoridor.gui.GUIConstants;
+import it.units.sdm.quoridor.gui.GameBoardGUI;
+import it.units.sdm.quoridor.gui.GameController;
+import it.units.sdm.quoridor.gui.buttons.CancelButton;
+import it.units.sdm.quoridor.gui.buttons.OrientationButton;
 import it.units.sdm.quoridor.utils.WallOrientation;
 
 import javax.swing.*;
